+++ packages/integration-tests/src/test-projects/conversation-handler/amplify/custom-conversation-handler/custom_handler.ts
@@ -1,24 +1,37 @@
 import {
   ConversationTurnEvent,
-  ExecutableTool,
-  ToolResultContentBlock,
+  createExecutableTool,
   handleConversationTurnEvent,
 } from '@aws-amplify/backend-ai/conversation/runtime';
-import { expectedTemperatureInProgrammaticToolScenario } from '../constants.js';
+import { expectedTemperaturesInProgrammaticToolScenario } from '../constants.js';
 
-const thermometer: ExecutableTool = {
-  name: 'thermometer',
-  description: 'Returns current temperature in Seattle',
-  execute: (): Promise<ToolResultContentBlock> => {
-    return Promise.resolve({
-      // We use this value in test assertion.
-      // LLM uses tool to get temperature and serves this value in final response.
-      // We're matching number only as LLM may translate unit to something more descriptive.
-      text: `${expectedTemperatureInProgrammaticToolScenario}F`,
-    });
+const thermometerInputSchema = {
+  type: 'object',
+  properties: {
+    city: { type: 'string' },
   },
-  inputSchema: { json: { type: 'object' } },
-};
+  required: ['city'],
+} as const;
+
+const thermometer = createExecutableTool(
+  'thermometer',
+  'Returns current temperature in cities',
+  {
+    json: thermometerInputSchema,
+  },
+  (input) => {
+    const city = input.city;
+    if (city === 'Seattle' || city === 'Boston' || city === 'Miami') {
+      return Promise.resolve({
+        // We use this value in test assertion.
+        // LLM uses tool to get temperature and serves this value in final response.
+        // We're matching number only as LLM may translate unit to something more descriptive.
+        text: `${expectedTemperaturesInProgrammaticToolScenario[city]}F`,
+      });
+    }
+    throw new Error(`Unknown city ${input.city}`);
+  }
+);
 
 /**
  * Handler with simple tool.