+++ scripts/components/api-changes-validator/test-resources/test-projects/without-breaks/project-without-breaks/src/index.ts
@@ -110,3 +110,8 @@ export type SampleTypeUsingClass = {
 export type SampleTypeThatReferencesFunction<T extends typeof someFunction1> = {
   sampleProperty: T;
 };
+
+// This type is intentionally different from what's in api report
+export type SampleIgnoredType = {
+  someProperty: number;
+};